react-select component 재사용 가능하게 분리. (width props) / tech array -> data 폴더로 이동
