+++ src/data/techListData.ts
@@ -1,0 +1,27 @@
+const options = [
+  { value: 'javascript', label: 'javascript' },
+  { value: 'java', label: 'java' },
+  { value: 'typescript', label: 'typescript' },
+  { value: 'react', label: 'react' },
+  { value: 'spring', label: 'spring' },
+  { value: 'c/c++', label: 'c/c++' },
+  { value: 'angular', label: 'angular' },
+  { value: 'vue', label: 'vue' },
+  { value: 'next', label: 'next' },
+  { value: 'nuxt', label: 'nuxt' },
+  { value: 'cloud', label: 'cloud' },
+  { value: 'CI/CD', label: 'CI/CD' },
+  { value: 'testing', label: 'testing' },
+  { value: 'flutter', label: 'flutter' },
+  { value: 'kotlin', label: 'kotlin' },
+  { value: 'android', label: 'android' },
+  { value: 'ios', label: 'ios' },
+  { value: 'c#', label: 'c#' },
+  { value: 'rust', label: 'rust' },
+  { value: 'nest', label: 'nest' },
+  { value: 'R', label: 'R' },
+  { value: 'python', label: 'python' },
+  { value: 'AI', label: 'AI' },
+];
+
+export default options;
